backend/requirements.txt
Django>=3.2,<4.3
djangorestframework>=3.12
django-cors-headers>=3.7
djangorestframework-simplejwt>=5.2
openpyxl>=3.0
django-filter>=2.4
drf-yasg>=1.20
google-api-python-client>=2.70
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.7.0
pyjwt>=2.6
python-dotenv>=1.0.0
gunicorn>=20.1.0
